fix(postgres): broader scan — pgx simple-protocol type issues
Audit-found two additional call sites with the same root cause as the
market_intelligence NullFloat64 error:

1. toNullString helper returned sql.NullString, used in
   dh_suggestions_repository.go to bind optional TEXT columns. Same
   encoding failure under QueryExecModeExec. Fix: delete the helper
   (the already-existing nullIfEmpty returns 'any' with nil/string,
   which pgx handles natively), migrate the four callsites.

2. Two UPDATE queries reused $1 in three contexts (SET assignment +
   two CASE WHEN comparisons), triggering SQLSTATE 42P08
   'inconsistent types deduced for parameter':
     - pricing_store.SetReviewedPrice
     - purchase_price_store (override price)
   Fix: explicit ::BIGINT casts pin the type consistently.

A third site (purchase_dh_store.UpdatePurchaseDHPushStatus) already
has a workaround of its own (passes the status arg twice as $1 and
$2) — leaving that as-is.

Named-string args (inventory.Phase, inventory.SnapshotStatus, etc.)
already have string() casts at every call site per the Phase 1 spike
learning. No action needed.

+++ internal/adapters/storage/postgres/pricing_store.go
@@ -35,11 +35,16 @@ func (prs *PricingStore) UpdateReviewedPrice(ctx context.Context, purchaseID str
 	}
 	// When a reviewed price is committed, any outstanding AI suggestion is
 	// superseded — clearing it keeps the "Needs Attention" filter honest.
+	// $1 is reused in three contexts (SET int, two CASE WHEN comparisons).
+	// pgx QueryExecModeExec uses extended protocol with type inference;
+	// reusing $1 in heterogenous contexts trips SQLSTATE 42P08
+	// ("inconsistent types deduced for parameter"). Explicit ::BIGINT casts
+	// pin the type consistently.
 	result, err := prs.db.ExecContext(ctx,
 		`UPDATE campaign_purchases
-		 SET reviewed_price_cents = $1, reviewed_at = $2, review_source = $3,
-		     ai_suggested_price_cents = CASE WHEN $1 > 0 THEN 0 ELSE ai_suggested_price_cents END,
-		     ai_suggested_at         = CASE WHEN $1 > 0 THEN '' ELSE ai_suggested_at END,
+		 SET reviewed_price_cents = $1::BIGINT, reviewed_at = $2, review_source = $3,
+		     ai_suggested_price_cents = CASE WHEN $1::BIGINT > 0 THEN 0 ELSE ai_suggested_price_cents END,
+		     ai_suggested_at         = CASE WHEN $1::BIGINT > 0 THEN '' ELSE ai_suggested_at END,
 		     updated_at = $4
 		 WHERE id = $5`,
 		priceCents, reviewedAt, source, now, purchaseID,
